Marked all none public classes as internal.
Removed the getFullPath from the asset, cause its not used.
Fixed getFullUrl.
Added the "print_string" (format strings) to the knownTypes array in the handler.
Fixed most of the print method.
Added get/set baseUrl to the container.
Still not sure that the container should handle this, but it works for now.
Added a get/set container method to the Asset class.
Updated tests acordingly.